requirements.txt
streamlit
Pillow
streamlit-image-comparison